auction_system:Implemented use cases 1.seller will register product 2.Buyer will place bids 3.End the auction and set winner

+++ src/main/java/com/auction/service/BidService.java
@@ -3,5 +3,5 @@ package com.auction.service;
 import com.auction.model.Bid;
 
 public interface BidService {
-    Bid placeBid(String token, long productId, double amount);
+    Bid placeBid(String token, long productId, long amount);
 }
